SF-178: Move My Account to XForge Common
- Move My Account and related Paratext items to xforge-common
- Remove Change Username Dialog
