+++ src/SIL.XForge.Scripture/ClientApp/src/app/connect-project/connect-project.component.ts
@@ -5,10 +5,10 @@ import { Router } from '@angular/router';
 import { InputSystem } from 'xforge-common/models/input-system';
 import { SubscriptionDisposable } from 'xforge-common/subscription-disposable';
 import { UserService } from 'xforge-common/user.service';
-import { ParatextProject } from '../core/models/paratext-project';
+import { ParatextProject } from '../../xforge-common/models/paratext-project';
+import { ParatextService } from '../../xforge-common/paratext.service';
 import { SFProject } from '../core/models/sfproject';
 import { SyncJob } from '../core/models/sync-job';
-import { ParatextService } from '../core/paratext.service';
 import { SFProjectUserService } from '../core/sfproject-user.service';
 import { SFProjectService } from '../core/sfproject.service';
 import { SyncJobService } from '../core/sync-job.service';
