+++ src/SIL.XForge.Scripture/ClientApp/src/app/sync/sync.component.ts
@@ -5,8 +5,8 @@ import { Observable } from 'rxjs';
 
 import { NoticeService } from 'xforge-common/notice.service';
 import { SubscriptionDisposable } from 'xforge-common/subscription-disposable';
+import { ParatextService } from '../../xforge-common/paratext.service';
 import { SyncJob } from '../core/models/sync-job';
-import { ParatextService } from '../core/paratext.service';
 import { SFProjectService } from '../core/sfproject.service';
 import { SyncJobService } from '../core/sync-job.service';
