+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/my-account/delete-account-dialog/delete-account-dialog.component.ts
