+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/my-account/my-account.component.html
