+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/paratext.service.spec.ts
