+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/xforge-common.module.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { ngfModule } from 'angular-file';
 import { OAuthModule } from 'angular-oauth2-oidc';
 import { AvatarModule } from 'ngx-avatar';
 
@@ -8,6 +9,8 @@ import { AvatarComponent } from './avatar/avatar.component';
 import { ChangePasswordComponent } from './change-password/change-password.component';
 import { EmailInviteComponent } from './email-invite/email-invite.component';
 import { InviteDialogComponent } from './email-invite/invite-dialog.component';
+import { DeleteAccountDialogComponent } from './my-account/delete-account-dialog/delete-account-dialog.component';
+import { MyAccountComponent } from './my-account/my-account.component';
 import { ProjectsComponent } from './projects/projects.component';
 import { SaDeleteDialogComponent } from './system-administration/sa-delete-dialog.component';
 import { SaUserEntryComponent } from './system-administration/sa-user-entry.component';
@@ -18,8 +21,10 @@ import { UICommonModule } from './ui-common.module';
 const componentExports = [
   AvatarComponent,
   ChangePasswordComponent,
+  DeleteAccountDialogComponent,
   EmailInviteComponent,
   InviteDialogComponent,
+  MyAccountComponent,
   ProjectsComponent,
   SaDeleteDialogComponent,
   SaUserEntryComponent,
@@ -34,6 +39,7 @@ export const xForgeCommonEntryComponents = [InviteDialogComponent, SaDeleteDialo
     // AvatarModule included here rather than `ui-common.module.ts` so unit tests don't access the internet
     AvatarModule,
     CommonModule,
+    ngfModule,
     OAuthModule.forRoot({
       resourceServer: {
         allowedUrls: ['json-api'],
@@ -44,6 +50,7 @@ export const xForgeCommonEntryComponents = [InviteDialogComponent, SaDeleteDialo
     UICommonModule
   ],
   declarations: componentExports,
-  exports: componentExports
+  exports: componentExports,
+  entryComponents: [DeleteAccountDialogComponent]
 })
 export class XForgeCommonModule {}
